+++ lib/models/message_content.dart
@@ -1,3 +1,9 @@
+import 'dart:ui';
+
+import 'package:flutter/cupertino.dart';
+
+import '../chat_analyzer/utilities/fix_dates_utilities.dart';
+
 /// Each message in the chat class
 class MessageContent {
   MessageContent({required this.senderId, required this.msg, this.dateTime});
@@ -6,7 +12,8 @@ class MessageContent {
   final String? msg;
   final DateTime? dateTime;
 
-  Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() =>
+      {
         'senderId': senderId,
         'msg': msg,
         'dateTime': dateTime.toString(),
@@ -15,6 +22,19 @@ class MessageContent {
   @override
   String toString() {
     return 'MessageContent{senderId: $senderId, msg: $msg, dateTime: $dateTime}';
+  }
+
+  bool isImage() {
+    RegExp pattern = RegExp(r'IMG-\d\d\d\d\d\d\d\d-WA\d\d\d\d\d?\d?[.]jpg');
+    if (dateTime != null && msg != null) {
+      String dateTimeString = dateTime!.year.toString() +
+          FixDateUtilities.fixMonthOrDayTo01(dateTime!.month.toString()) +
+          FixDateUtilities.fixMonthOrDayTo01(dateTime!.day.toString());
+      return msg!.startsWith("IMG-$dateTimeString-WA") && msg!.startsWith(pattern);
+    } else if (msg != null) {
+      return msg!.startsWith(pattern);
+    }
+    return false;
   }
 
   static MessageContent fromJson(Map<String, dynamic> messageMap) {
@@ -42,15 +62,15 @@ class MessageContent {
     return msgsParsed;
   }
 
-  // opreator ==
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is MessageContent &&
-          runtimeType == other.runtimeType &&
-          senderId == other.senderId &&
-          msg == other.msg &&
-          dateTime == other.dateTime;
+          other is MessageContent &&
+              runtimeType == other.runtimeType &&
+              senderId == other.senderId &&
+              msg == other.msg &&
+              dateTime == other.dateTime;
 
-
+  @override
+  int get hashCode => hashValues(senderId, msg, dateTime);
 }